OSSM-2170 Ensure ServiceMeshMembers are deleted when SMMR is cluster-scoped (#1048)
This also fixes edge cases where the SMMR.spec.members list didn't reflect the actual SMM objects. For example, when the user removes namespaces from the SMMR.spec.members and then immediately deletes the SMMR. In that case, the controller didn't delete the SMM when finalizing the SMMR, because the SMMR.spec.members list was empty.

+++ pkg/controller/servicemesh/memberroll/controller.go
@@ -555,16 +555,19 @@ func (r *MemberRollReconciler) ensureMemberExists(ctx context.Context, ns, meshN
 }
 
 func (r *MemberRollReconciler) finalizeObject(ctx context.Context, roll *maistrav1.ServiceMeshMemberRoll) (ok bool, err error) {
-	for _, ns := range roll.Spec.Members {
-		member := maistrav1.ServiceMeshMember{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      common.MemberName,
-				Namespace: ns,
-			},
+	members := &maistrav1.ServiceMeshMemberList{}
+	err = r.Client.List(ctx, members, client.MatchingFields{"spec.controlPlaneRef.namespace": roll.Namespace})
+	if err != nil {
+		return false, err
+	}
+
+	for _, member := range members.Items {
+		if member.DeletionTimestamp != nil || !isCreatedByThisController(&member) {
+			continue
 		}
 		err := r.Client.Delete(ctx, &member)
 		if err != nil && !errors.IsNotFound(err) {
-			return false, pkgerrors.Wrapf(err, "Could not delete ServiceMeshMember %s/%s", ns, common.MemberName)
+			return false, pkgerrors.Wrapf(err, "Could not delete ServiceMeshMember %s/%s", member.Namespace, member.Name)
 		}
 	}
